Rename 'MemberOfToolbar' to 'MemberOfToolbarOld'
The `MemberOfToolbar` component
should be renamed to `MemberOfToolbarOld`
because a new version of the component will
be used instead.
