+++ src/pages/ActiveUsers/UserMemberOf.tsx
@@ -11,7 +11,7 @@ import {
   PaginationVariant,
 } from "@patternfly/react-core";
 // Others
-import MemberOfToolbar from "src/components/MemberOf/MemberOfToolbar";
+import MemberOfToolbar from "src/components/MemberOf/MemberOfToolbarOld";
 import MemberOfTable from "src/components/MemberOf/MemberOfTable";
 // Data types
 import {
